Add initial attemps at project root detection and compilation
Current problem:
Open root project file
Use Vim's `:edit` to open a "child" file, which specifies the project
root as e.g. "../../root.tex" (relative to the child file).
Problem is that the relative "../../" acts relative to the first file
opened in Vim, and not to the currently edited file.